python/requirements.txt
torch
torchvision
tqdm
Pillow
